Movement of character type to own file and redoing imports
