+++ src/components/CharacterList/CharacterListTypes.ts
@@ -1,0 +1,7 @@
+export interface Character {
+  name: string;
+  category: string;
+  description: string;
+  significanceIndex: number;
+  avatar: string;
+}
